Fix save confirmation message to display in message area
Changed from using beginbottom()/endbottom() to using message() function
which displays in the message area at the top of the screen. This matches
the pattern used for other success messages and ensures the confirmation
is visible to the user.

Now when a file is saved, the user will see:
  'File saved to: /full/path/to/file.lgf'
in the message area, similar to how other messages are displayed.